@@ -36,6 +36,6 @@
     "@vercel/node": "^2.8.15",
     "eslint": "^8.3.0",
     "jest": "^29.4.3",
-    "typescript": "^4.8.4"
+    "typescript": "^5.1.6"
   }
 }
