@@ -1,4 +1,4 @@
-import { Session } from "@supabase/gotrue-js";
+import { Session } from "@supabase/supabase-js";
 import {
   createContext,
   Dispatch,
@@ -8,7 +8,6 @@ import {
   useContext,
   useEffect,
   useMemo,
-  useRef,
   useState,
 } from "react";
 import { useLocation } from "react-router-dom";
@@ -16,12 +15,10 @@ import useLocalStorage from "react-use-localstorage";
 import Stripe from "stripe";
 
 import { LOCAL_STORAGE_SETTINGS_KEY } from "../lib/constants";
-import { loadSponsorOnlyLayouts } from "../lib/cytoscape";
 import { useCustomerInfo, useHostedCharts } from "../lib/queries";
-import { supabase } from "../lib/supabaseClient";
-import { useGraphStore } from "../lib/useGraphStore";
 import { languages } from "../locales/i18n";
 import { colors, darkTheme } from "../slang/config";
+import { supabase } from "../lib/supabaseClient";
 
 type Theme = typeof colors;
 
@@ -114,38 +111,26 @@ const Provider = ({ children }: { children?: ReactNode }) => {
 
   const [checkedSession, setCheckedSession] = useState(false);
   const [session, setSession] = useState<Session | null>(null);
-  const sponsorLayoutsLoading = useRef(false);
-
-  /* Load Sponsor-only layouts when logged in */
-  useEffect(() => {
-    // If not logged in, return
-    if (!session) return;
-    // If already loaded, return
-    if (useGraphStore.getState().sponsorLayoutsLoaded) return;
-    // If in the process of loading, return
-    if (sponsorLayoutsLoading.current) return;
-    sponsorLayoutsLoading.current = true;
-    loadSponsorOnlyLayouts().then(() => {
-      useGraphStore.setState({ sponsorLayoutsLoaded: true });
-    });
-  }, [session]);
 
   useEffect(() => {
-    if (!supabase) {
-      setCheckedSession(true);
-      return;
-    }
+    requestAnimationFrame(() => {
+      if (!supabase) {
+        setCheckedSession(true);
+        return;
+      }
 
-    (async () => {
-      const {
-        data: { session },
-      } = await supabase.auth.getSession();
-      setSession(session);
-      setCheckedSession(true);
-      supabase.auth.onAuthStateChange((_event, session) => {
+      (async () => {
+        const {
+          data: { session },
+        } = await supabase.auth.getSession();
+        console.log("Resolve session");
         setSession(session);
-      });
-    })();
+        setCheckedSession(true);
+        supabase.auth.onAuthStateChange((_event, session) => {
+          setSession(session);
+        });
+      })();
+    });
   }, []);
 
   // Close Share Modal when navigating
