@@ -7,14 +7,18 @@ import { AppContext } from "./AppContext";
 export default function ColorMode() {
   const { theme } = useContext(AppContext);
 
-  return createPortal(
-    <style
-      dangerouslySetInnerHTML={{
-        __html: `:root {${Object.entries(theme)
-          .map(([key, value]) => `--color-${key}: ${value};`)
-          .join("\n")}}`,
-      }}
-    />,
-    document.head
+  return (
+    <>
+      {createPortal(
+        <style
+          dangerouslySetInnerHTML={{
+            __html: `:root {${Object.entries(theme)
+              .map(([key, value]) => `--color-${key}: ${value};`)
+              .join("\n")}}`,
+          }}
+        />,
+        document.head
+      )}
+    </>
   );
 }
