@@ -1,5 +1,5 @@
 import { useContext, useMemo } from "react";
-import { useLocation, useRouteMatch } from "react-router-dom";
+import { useLocation, useParams } from "react-router-dom";
 
 import { AppContext } from "../components/AppContext";
 import { slugify } from "./helpers";
@@ -16,29 +16,27 @@ export function getAnimationSettings() {
 }
 
 export function useFullscreen() {
+  const params = useParams();
   const { pathname } = useLocation();
-  const { path } = useRouteMatch();
-  return pathname === "/f" || path === "/p/:public_id";
+  return pathname === "/f" || "public_id" in params;
 }
 
 export function useIsReadOnly() {
-  const { path } = useRouteMatch();
+  const { pathname } = useLocation();
+  const params = useParams();
   return (
-    path === "/p/:public_id" ||
-    path === "/f" ||
-    path === "/c/:graphText?" ||
-    path === "/r/:graphText?"
+    "public_id" in params || ["/f", "/c", "/r"].some((k) => k === pathname)
   );
 }
 
 export function useIsLocalChart() {
-  const { path } = useRouteMatch();
-  return useMemo(() => path === "/:workspace" || path === "/", [path]);
+  const { pathname } = useLocation();
+  return useMemo(() => pathname === "/", [pathname]);
 }
 
 export function useIsPublicHostedCharted() {
-  const { path } = useRouteMatch();
-  return useMemo(() => path === "/p/:public_id", [path]);
+  const params = useParams();
+  return useMemo(() => "public_id" in params, [params]);
 }
 
 /**
@@ -88,14 +86,14 @@ export function useIsFirefox() {
  * It's used to alter CSS with data-showing
  */
 export function useIsEditorView() {
-  const { path, isExact } = useRouteMatch();
+  const pathname = useLocation().pathname;
+  const params = useParams();
+
   return (
-    path === "/u/:id" ||
-    path === "/c/:graphText?" ||
-    path === "/r/:graphText?" ||
-    path === "/h" ||
-    path === "/:workspace" ||
-    (path === "/" && isExact)
+    // works for hosted and local charts
+    ["id", "workspace"].some((k) => k in params) ||
+    // works for read-only charts that aren't fullscreen
+    ["/c", "/r", "/"].some((k) => pathname === k)
   );
 }
 
